requirements.txt
numpy==2.2
pandas>=2.3
scikit-learn>=1.7
lightgbm>=4.6
xgboost>=3.0
optuna>=4.2
matplotlib>=3.10
seaborn>=0.13.2
pyyaml
joblib>=1.2
pybaseball==2.2.7
psutil>=5.9
threadpoolctl>=3.1.0
scipy>=1.14.0
pyarrow>=17.0.0
requests==2.32.3
lxml==5.2.2
beautifulsoup4==4.12.3
xxhash==3.4.1
hdbscan==0.8.40
numba==0.61.2
llvmlite==0.44.0
umap-learn>=0.5.6
shap>=0.45.0
polars>=1.31
nfl_data_py>=0.3.1
